update refactoring, update documentation, remove unecessary example, update tests
