+++ python/packages/autogen-ext/src/autogen_ext/memory/chromadb/__init__.py
@@ -1,0 +1,22 @@
+from ._chromadb import ChromaDBVectorMemory
+from ._chroma_configs import (
+    ChromaDBVectorMemoryConfig,
+    CustomEmbeddingFunctionConfig,
+    DefaultEmbeddingFunctionConfig,
+    HttpChromaDBVectorMemoryConfig,
+    OpenAIEmbeddingFunctionConfig,
+    PersistentChromaDBVectorMemoryConfig,
+    SentenceTransformerEmbeddingFunctionConfig,
+)
+
+
+__all__ = [
+    "ChromaDBVectorMemory",
+    "ChromaDBVectorMemoryConfig",
+    "PersistentChromaDBVectorMemoryConfig",
+    "HttpChromaDBVectorMemoryConfig",
+    "DefaultEmbeddingFunctionConfig",
+    "SentenceTransformerEmbeddingFunctionConfig",
+    "OpenAIEmbeddingFunctionConfig",
+    "CustomEmbeddingFunctionConfig",
+]
