+++ python/packages/autogen-ext/src/autogen_ext/memory/chromadb/_chromadb.py
@@ -1,6 +1,6 @@
 import logging
 import uuid
-from typing import Any, Callable, Dict, List, Literal, cast
+from typing import Any, List
 
 from autogen_core import CancellationToken, Component, Image
 from autogen_core.memory import Memory, MemoryContent, MemoryMimeType, MemoryQueryResult, UpdateContextResult
@@ -9,8 +9,17 @@ from autogen_core.models import SystemMessage
 from chromadb import HttpClient, PersistentClient
 from chromadb.api.models.Collection import Collection
 from chromadb.api.types import Document, Metadata
-from pydantic import BaseModel, Field
 from typing_extensions import Self
+
+from ._chroma_configs import (
+    ChromaDBVectorMemoryConfig,
+    CustomEmbeddingFunctionConfig,
+    DefaultEmbeddingFunctionConfig,
+    HttpChromaDBVectorMemoryConfig,
+    OpenAIEmbeddingFunctionConfig,
+    PersistentChromaDBVectorMemoryConfig,
+    SentenceTransformerEmbeddingFunctionConfig,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -21,146 +30,6 @@ except ImportError as e:
     raise ImportError(
         "To use the ChromaDBVectorMemory the chromadb extra must be installed. Run `pip install autogen-ext[chromadb]`"
     ) from e
-
-
-class BaseEmbeddingFunctionConfig(BaseModel):
-    """Base configuration for embedding functions.
-
-    .. versionadded:: v0.4.1
-       Support for custom embedding functions in ChromaDB memory.
-    """
-
-    function_type: Literal["default", "sentence_transformer", "openai", "custom"]
-
-
-class DefaultEmbeddingFunctionConfig(BaseEmbeddingFunctionConfig):
-    """Configuration for the default ChromaDB embedding function.
-
-    Uses ChromaDB's default embedding function (Sentence Transformers all-MiniLM-L6-v2).
-
-    .. versionadded:: v0.4.1
-       Support for custom embedding functions in ChromaDB memory.
-    """
-
-    function_type: Literal["default"] = "default"
-
-
-class SentenceTransformerEmbeddingFunctionConfig(BaseEmbeddingFunctionConfig):
-    """Configuration for SentenceTransformer embedding functions.
-
-    Allows specifying a custom SentenceTransformer model for embeddings.
-
-    .. versionadded:: v0.4.1
-       Support for custom embedding functions in ChromaDB memory.
-
-    Args:
-        model_name (str): Name of the SentenceTransformer model to use.
-            Defaults to "all-MiniLM-L6-v2".
-
-    Example:
-        .. code-block:: python
-
-            config = SentenceTransformerEmbeddingFunctionConfig(model_name="paraphrase-multilingual-mpnet-base-v2")
-    """
-
-    function_type: Literal["sentence_transformer"] = "sentence_transformer"
-    model_name: str = Field(default="all-MiniLM-L6-v2", description="SentenceTransformer model name to use")
-
-
-class OpenAIEmbeddingFunctionConfig(BaseEmbeddingFunctionConfig):
-    """Configuration for OpenAI embedding functions.
-
-    Uses OpenAI's embedding API for generating embeddings.
-
-    .. versionadded:: v0.4.1
-       Support for custom embedding functions in ChromaDB memory.
-
-    Args:
-        api_key (str): OpenAI API key. If empty, will attempt to use environment variable.
-        model_name (str): OpenAI embedding model name. Defaults to "text-embedding-ada-002".
-
-    Example:
-        .. code-block:: python
-
-            config = OpenAIEmbeddingFunctionConfig(api_key="sk-...", model_name="text-embedding-3-small")
-    """
-
-    function_type: Literal["openai"] = "openai"
-    api_key: str = Field(default="", description="OpenAI API key")
-    model_name: str = Field(default="text-embedding-ada-002", description="OpenAI embedding model name")
-
-
-class CustomEmbeddingFunctionConfig(BaseEmbeddingFunctionConfig):
-    """Configuration for custom embedding functions.
-
-    Allows using a custom function that returns a ChromaDB-compatible embedding function.
-
-    .. versionadded:: v0.4.1
-       Support for custom embedding functions in ChromaDB memory.
-
-    .. warning::
-       Configurations containing custom functions are not serializable.
-
-    Args:
-        function (Callable): Function that returns a ChromaDB-compatible embedding function.
-        params (Dict[str, Any]): Parameters to pass to the function.
-
-    Example:
-        .. code-block:: python
-
-            def create_my_embedder(param1="default"):
-                # Return a ChromaDB-compatible embedding function
-                class MyCustomEmbeddingFunction(EmbeddingFunction):
-                    def __call__(self, input: Documents) -> Embeddings:
-                        # Custom embedding logic here
-                        return embeddings
-
-                return MyCustomEmbeddingFunction(param1)
-
-
-            config = CustomEmbeddingFunctionConfig(function=create_my_embedder, params={"param1": "custom_value"})
-    """
-
-    function_type: Literal["custom"] = "custom"
-    function: Callable[..., Any] = Field(description="Function that returns an embedding function")
-    params: Dict[str, Any] = Field(default_factory=dict, description="Parameters to pass to the function")
-
-
-class ChromaDBVectorMemoryConfig(BaseModel):
-    """Base configuration for ChromaDB-based memory implementation.
-
-    .. versionchanged:: v0.4.1
-       Added support for custom embedding functions via embedding_function_config.
-    """
-
-    client_type: Literal["persistent", "http"]
-    collection_name: str = Field(default="memory_store", description="Name of the ChromaDB collection")
-    distance_metric: str = Field(default="cosine", description="Distance metric for similarity search")
-    k: int = Field(default=3, description="Number of results to return in queries")
-    score_threshold: float | None = Field(default=None, description="Minimum similarity score threshold")
-    allow_reset: bool = Field(default=False, description="Whether to allow resetting the ChromaDB client")
-    tenant: str = Field(default="default_tenant", description="Tenant to use")
-    database: str = Field(default="default_database", description="Database to use")
-    embedding_function_config: BaseEmbeddingFunctionConfig = Field(
-        default_factory=DefaultEmbeddingFunctionConfig, description="Configuration for the embedding function"
-    )
-
-
-class PersistentChromaDBVectorMemoryConfig(ChromaDBVectorMemoryConfig):
-    """Configuration for persistent ChromaDB memory."""
-
-    client_type: Literal["persistent", "http"] = "persistent"
-    persistence_path: str = Field(default="./chroma_db", description="Path for persistent storage")
-
-
-class HttpChromaDBVectorMemoryConfig(ChromaDBVectorMemoryConfig):
-    """Configuration for HTTP ChromaDB memory."""
-
-    client_type: Literal["persistent", "http"] = "http"
-    host: str = Field(default="localhost", description="Host of the remote server")
-    port: int = Field(default=8000, description="Port of the remote server")
-    ssl: bool = Field(default=False, description="Whether to use HTTPS")
-    headers: Dict[str, str] | None = Field(default=None, description="Headers to send to the server")
 
 
 class ChromaDBVectorMemory(Memory, Component[ChromaDBVectorMemoryConfig]):
@@ -294,38 +163,35 @@ class ChromaDBVectorMemory(Memory, Component[ChromaDBVectorMemoryConfig]):
 
         config = self._config.embedding_function_config
 
-        if config.function_type == "default":
+        if isinstance(config, DefaultEmbeddingFunctionConfig):
             return embedding_functions.DefaultEmbeddingFunction()
 
-        elif config.function_type == "sentence_transformer":
-            cfg = cast(SentenceTransformerEmbeddingFunctionConfig, config)
+        elif isinstance(config, SentenceTransformerEmbeddingFunctionConfig):
             try:
-                return embedding_functions.SentenceTransformerEmbeddingFunction(model_name=cfg.model_name)
+                return embedding_functions.SentenceTransformerEmbeddingFunction(model_name=config.model_name)
             except Exception as e:
                 raise ImportError(
-                    f"Failed to create SentenceTransformer embedding function with model '{cfg.model_name}'. "
+                    f"Failed to create SentenceTransformer embedding function with model '{config.model_name}'. "
                     f"Ensure sentence-transformers is installed and the model is available. Error: {e}"
                 ) from e
 
-        elif config.function_type == "openai":
-            cfg = cast(OpenAIEmbeddingFunctionConfig, config)
+        elif isinstance(config, OpenAIEmbeddingFunctionConfig):
             try:
-                return embedding_functions.OpenAIEmbeddingFunction(api_key=cfg.api_key, model_name=cfg.model_name)
+                return embedding_functions.OpenAIEmbeddingFunction(api_key=config.api_key, model_name=config.model_name)
             except Exception as e:
                 raise ImportError(
-                    f"Failed to create OpenAI embedding function with model '{cfg.model_name}'. "
+                    f"Failed to create OpenAI embedding function with model '{config.model_name}'. "
                     f"Ensure openai is installed and API key is valid. Error: {e}"
                 ) from e
 
-        elif config.function_type == "custom":
-            cfg = cast(CustomEmbeddingFunctionConfig, config)
+        elif isinstance(config, CustomEmbeddingFunctionConfig):
             try:
-                return cfg.function(**cfg.params)
+                return config.function(**config.params)
             except Exception as e:
                 raise ValueError(f"Failed to create custom embedding function. Error: {e}") from e
 
         else:
-            raise ValueError(f"Unsupported embedding function type: {config.function_type}")
+            raise ValueError(f"Unsupported embedding function config type: {type(config)}")
 
     def _ensure_initialized(self) -> None:
         """Ensure ChromaDB client and collection are initialized."""
